+++ packages/devui-vue/devui/menu/src/components/sub-menu.tsx
@@ -9,7 +9,7 @@ import {
   watchEffect,
   watch,
 } from 'vue';
-import { addLayer, pushElement, clearSelect,getLayer } from '../composables/layer-composables';
+import { addLayer, pushElement, clearSelect,getLayer } from '../helper/layer-composables';
 import { useClick } from '../hook/use-click';
 import { useShowSubMenu } from '../hook/use-show-submenu';
 import { SubMenuProps, subMenuProps } from '../types/sub-menu-types';
